Addressed numbers maintenance, user info, feature reqquests, and bugs
Maintenance: #193, #208, User Info: #195, #204. Feature: #203, #211, #212, #214, Bug: #210, #217, #218, #219, #220, #222

+++ NetworkModule/Spi.c
@@ -50,7 +50,7 @@
 
 void spi_init(void)
 {
-  uint8_t i;
+  int i;
   
   // Set specific GPIO pins to initial values for use as SPI bit
   // bang. Hold ENC28J60 in reset until SPI pins are set up.
